pyproject.toml
[tool.poetry]
name = "vam_backend"
version = "0.1.0"
description = "Deps for the VAM backend alpha.0.1"
authors = ["NFC Solutions"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "~3.2"
python-decouple = "^3.5"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"
django-cors-headers = "^3.10.0"
serializers = "^0.2.4"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
